Delegate Package logic to middleware
